+++ examples/sdlevent.py
@@ -19,11 +19,11 @@ def main():
                     raise SystemExit()
                 elif event.type == "MOUSEMOTION":
                     console.ch[:, -1] = 0
-                    console.print_(0, HEIGHT - 1, repr(event))
+                    console.print_(0, HEIGHT - 1, str(event))
                 else:
                     console.blit(console, 0, 0, 0, 1, WIDTH, HEIGHT - 2)
                     console.ch[:, -3] = 0
-                    console.print_(0, HEIGHT - 3, repr(event))
+                    console.print_(0, HEIGHT - 3, str(event))
 
 
 if __name__ == "__main__":